working version

+++ processor.py
@@ -6,6 +6,8 @@ import logging
 
 from utils import FileUtils
 from converters import Converter, ConversionStrategy
+from Album import Album
+from Song import Song
 
 logger = logging.getLogger(__name__)
 
@@ -14,6 +16,15 @@ class FileProcessor:
     def __init__(self, library_directory: str, strategy: ConversionStrategy):
         self.dir_library = library_directory
         self.converter = Converter(strategy)
+        while True:
+            destructive = input("Will you want to delete the files after the process? type 1 if yes, otherwise type any other thing\n")
+            if destructive is 1:
+                self.destructive = True
+                break
+            else:
+                self.destructive = False
+                break
+        self.destructive = destructive
 
     @staticmethod
     def extract_flac_metadata(file_path: str) -> Dict[str, Any]:
@@ -25,9 +36,9 @@ class FileProcessor:
             metadata['file'] = file_path
             metadata['title'] = FileUtils.get_metadata(audio, 'title')
             metadata['album'] = FileUtils.get_metadata(audio, 'album')
-            metadata['artist'] = FileUtils.get_metadata(audio, 'artist')
+            metadata['artist'] = FileUtils.get_metadata(audio, 'albumartist')
             metadata['date'] = FileUtils.get_metadata(audio, 'date')
-            metadata['track_number'] = FileUtils.get_metadata(audio, 'track_number')
+            metadata['track_number'] = FileUtils.get_metadata(audio, 'tracknumber')
         except Exception as e:
             logger.error(f"Metadata extraction error for {file_path}: {e}")
             metadata['error'] = str(e)
@@ -35,58 +46,33 @@ class FileProcessor:
 
     def create_directory_structure(self, artist: str, album: str, date: str) -> str:
         """Create directory structure based on metadata."""
-        album_dir = f"{album} ({date})"
-        target_dir = os.path.join(self.dir_library, artist, album_dir)
+        album_dir = FileUtils.sanitize_directory_name(f"{artist.upper()}\\{album} ({date})")
+        target_dir = os.path.join(self.dir_library, album_dir)
         os.makedirs(target_dir, exist_ok=True)
         logger.info(f"Created directory: {target_dir}")
-        return target_dir
-
-    def get_flac_files_info(self, directory: str) -> List[Dict[str, Any]]:
-        """Retrieve and process metadata for all FLAC files in the directory."""
-        flac_metadata_list = []
-        for filename in os.listdir(directory):
-            if filename.lower().endswith('.flac'):
-                file_path = os.path.join(directory, filename)
-                FileUtils.remove_cover_image(file_path)
-                metadata = self.extract_flac_metadata(file_path)
-                flac_metadata_list.append(metadata)
-        return flac_metadata_list
-
-    def input_directory(self) -> str:
-        """Prompt user for input directory and validate it."""
-        directory = input("Please enter the source directory you want to work with\n")
-        while not os.path.isdir(directory) or not FileUtils.is_windows_directory_format(directory):
-            logger.warning(f"Invalid directory entered: {directory}")
-            directory = input("PLEASE ENTER A VALID DIRECTORY\n")
-        print(f'You selected the directory {directory}\n')
-        logger.info(f"Selected directory: {directory}")
-        return directory
-
-    def process_files(self) -> None:
-        """Process FLAC files in the given directory."""
-        logger.info("Starting file processing")
-        directory = self.input_directory()
-        flac_metadata_list = self.get_flac_files_info(directory)
-
-        if not flac_metadata_list or any('error' in metadata for metadata in flac_metadata_list):
-            # If any metadata entry is not valid, we ask for manual intervention
-            logger.warning("No FLAC files found or metadata extraction failed for all files.")
-            print("Metadata not found, please enter the metadata manually\n")
-            artist = input("Artist: \n")
-            album = input("Album Title: \n")
-            date = input("Date: \n")
-        else:
-            # All metadata entries are valid at this point
-            first_metadata = flac_metadata_list[0]
-            artist = first_metadata.get('artist', 'Unknown Artist')
-            album = first_metadata.get('album', 'Unknown Album')
-            date = first_metadata.get('date', 'Unknown Date')
-            logger.info(f"Metadata set from first file: Artist: {artist}, Album: {album}, Date: {date}")
-
-        target_dir = self.create_directory_structure(artist, album, date)
         print(f"Files will be processed into: {target_dir}")
 
-        for metadata in tqdm(flac_metadata_list, desc="Processing files", unit="file"):
+        return target_dir
+
+    def get_flac_files_info(self, album: Album):
+        """Retrieve and process metadata for all FLAC files in the directory."""
+        for filename in os.listdir(album.folder):
+            if filename.lower().endswith('.flac'):
+                file_path = os.path.join(album.folder, filename)
+                FileUtils.remove_cover_image(file_path)
+                metadata = self.extract_flac_metadata(file_path)
+                album.add_song(Song(file_path, metadata, album))
+
+    def process_files(self, album: Album) -> None:
+        """Process FLAC files in the given directory."""
+        logger.info("Starting file processing")
+        self.get_flac_files_info(album)
+        album.find_metadata()
+        target_dir = self.create_directory_structure(album.metadata['artist'], album.metadata['album'],
+                                                     album.metadata['date'])
+
+        for song in tqdm(album.songs, desc="Processing files", unit="file"):
+            metadata = song.metadata
             if 'error' in metadata:
                 logger.error(f"Error processing {metadata['file']}: {metadata['error']}")
                 print("Error processing {}: {}".format(metadata['file'], metadata['error']))
@@ -97,12 +83,11 @@ class FileProcessor:
 
             target_base_file_name = f"{track_number}. {title}"
             target_base_file_path = os.path.join(target_dir, target_base_file_name)
-
             self.converter.convert(metadata['file'], target_base_file_path)
-            logger.info(f"Removing original FLAC file: {metadata['file']}")
-            os.remove(metadata['file'])
+
+            if self.destructive:
+                logger.info(f"Removing original FLAC file: {metadata['file']}")
+                os.remove(metadata['file'])
 
         print("Processing completed.")
         logger.info("Processing completed.")
-
-
